examples/python/requirements.txt
aws-cdk-lib==2.212.0
constructs==10.0.5
ses-cloudwatch
